audio_test/backend/requirements.txt
fastapi
uvicorn[standard]
websockets
python-dotenv
dashscope
pydantic
python-multipart
